requirements.txt
websockets
confluent-kafka

requests
google-cloud-storage

apache-airflow-providers-snowflake
apache-airflow-providers-amazon


